+++ inlang/source-code/paraglide/paraglide-js/src/cli/steps/check-for-uncomitted-changes.ts
@@ -1,5 +1,5 @@
 import type { Logger } from "~/services/logger/index.js"
-import type { CliStep } from "../cli-utils.js"
+import type { CliStep } from "../utils.js"
 import { execAsync, prompt } from "~/cli/commands/init/utils.js"
 
 export const checkForUncommittedChanges: CliStep<{ logger: Logger }, unknown> = async (ctx) => {
