+++ inlang/source-code/paraglide/paraglide-js/src/cli/steps/prompt-for-outdir.ts
@@ -1,7 +1,7 @@
 import type { Logger } from "~/services/logger/index.js"
 import { DEFAULT_OUTDIR } from "~/cli/commands/init/defaults.js"
 import { prompt } from "~/cli/commands/init/utils.js"
-import type { CliStep } from "../cli-utils.js"
+import type { CliStep } from "../utils.js"
 
 export const promptForOutdir: CliStep<
 	{
