+++ inlang/source-code/paraglide/paraglide-js/src/cli/steps/update-ts-config.ts
@@ -1,5 +1,5 @@
 import type { Repository } from "@lix-js/client"
-import type { CliStep } from "../cli-utils.js"
+import type { CliStep } from "../utils.js"
 import type { Logger } from "~/services/logger/index.js"
 import { prompt } from "~/cli/commands/init/utils.js"
 import JSON5 from "json5"
